# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/summary:
    get:
      operationId: getTablesSummary
      tags: [tables]
      summary: Get table status summary counts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Table status counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TablesSummary"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a table by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update a table
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableBody"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Menu Items
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: tableId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/active:
    get:
      operationId: listActiveOrders
      tags: [orders]
      summary: Get active orders for kitchen view
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status or items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Payments
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: method
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment / complete a bill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/daily-report:
    get:
      operationId: getDailyReport
      tags: [payments]
      summary: Get daily sales report
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /payments/monthly-chart:
    get:
      operationId: getMonthlyChart
      tags: [payments]
      summary: Get monthly revenue chart data
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyRevenue"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Inventory
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, cooking, ready, billing]
        currentOrderId:
          type: ["integer", "null"]
      required:
        - id
        - number
        - capacity
        - status
    CreateTableBody:
      type: object
      properties:
        number:
          type: integer
        capacity:
          type: integer
      required:
        - number
        - capacity
    UpdateTableBody:
      type: object
      properties:
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, cooking, ready, billing]
    TablesSummary:
      type: object
      properties:
        available:
          type: integer
        occupied:
          type: integer
        cooking:
          type: integer
        ready:
          type: integer
        billing:
          type: integer
        total:
          type: integer
      required:
        - available
        - occupied
        - cooking
        - ready
        - billing
        - total

    # Menu Item
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemPrice:
          type: number
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
      required:
        - id
        - orderId
        - menuItemId
        - menuItemName
        - menuItemPrice
        - quantity
        - subtotal
    Order:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableNumber:
          type: integer
        waiterId:
          type: integer
        waiterName:
          type: string
        status:
          type: string
          enum: [pending, preparing, ready, billing, completed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tableId
        - tableNumber
        - waiterId
        - waiterName
        - status
        - items
        - totalAmount
        - createdAt
        - updatedAt
    CreateOrderItemInput:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string
      required:
        - menuItemId
        - quantity
    CreateOrderBody:
      type: object
      properties:
        tableId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemInput"
        notes:
          type: string
      required:
        - tableId
        - items
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, billing, completed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemInput"
        notes:
          type: string

    # Payment
    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        tableNumber:
          type: integer
        subtotal:
          type: number
        tax:
          type: number
        discount:
          type: number
        totalAmount:
          type: number
        method:
          type: string
          enum: [cash, upi, card]
        status:
          type: string
          enum: [pending, completed]
        createdAt:
          type: string
      required:
        - id
        - orderId
        - tableNumber
        - subtotal
        - tax
        - discount
        - totalAmount
        - method
        - status
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        discount:
          type: number
        method:
          type: string
          enum: [cash, upi, card]
        taxRate:
          type: number
      required:
        - orderId
        - method
    DailyReport:
      type: object
      properties:
        date:
          type: string
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalCustomers:
          type: integer
        cashRevenue:
          type: number
        upiRevenue:
          type: number
        cardRevenue:
          type: number
        avgOrderValue:
          type: number
      required:
        - date
        - totalRevenue
        - totalOrders
        - totalCustomers
        - cashRevenue
        - upiRevenue
        - cardRevenue
        - avgOrderValue
    DailyRevenue:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders

    # Inventory
...